test: GoodsItems 스토리 작성
RankingGoodsItem.stories.ts에 있던 내용 병합
